8、PH和温度范围的调节，要显示1位小数；(现在是PH值两位小数，温度只有1位)

+++ app/src/main/java/com/mibo/fishtank/activity/DeviceParamsActivity.java
@@ -132,12 +132,12 @@ public class DeviceParamsActivity extends BaseActivity {
             }
         });
         rangeTemp = (RangeSeekBar) findViewById(R.id.range_temp);
-        rangeTemp.setRangeValues(0, 40);
+        rangeTemp.setRangeValues(0f, 40f);
         rangeTemp.setOnRangeSeekBarChangeListener(new RangeSeekBar.OnRangeSeekBarChangeListener() {
             @Override
             public void onRangeSeekBarValuesChanged(RangeSeekBar bar, Object minValue, Object maxValue) {
-                deviceParams.TempMin = (int) minValue;
-                deviceParams.TempMax = (int) maxValue;
+                deviceParams.TempMin = (float) minValue;
+                deviceParams.TempMax = (float) maxValue;
             }
         });
